--- src/components/dashboard/services/referral/ReferralsBoard.tsx
@@ -1,10 +1,0 @@
-import ReferralsCard from "@/components/dashboard/services/referral/ReferralsCard.tsx";
-
-export default function ReferralsBoard() {
-
-    return (
-        <>
-            <ReferralsCard />
-        </>
-    );
-}
