+++ src/pages/LinktreePage.tsx
@@ -1,4 +1,4 @@
-import LandingPageNavBar from "@/components/landing/LandingPageNavBar";
+import LandingNavBar from "@/components/landing/LandingNavBar.tsx";
 import WorkInProgress from "@/components/WorkInProgress";
 import { OWNER_USERNAME } from "@/constants";
 import { useParams } from "react-router-dom";
@@ -8,7 +8,7 @@ export default function LinktreePage() {
 
   return (
       <>
-        {username == OWNER_USERNAME ? <LandingPageNavBar /> : null}
+        {username == OWNER_USERNAME ? <LandingNavBar /> : null}
         <WorkInProgress />
       </>
   )
